requirements/common.txt
Django==3.2.7
Pillow==8.3.2
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.8.6
django-cors-headers==3.9.0
django-filter==21.1
